Add back YearForm
- Combined CreateYear1 and yearForm into YearForm.js then deleted some of the older files
- added back year editing functionality using the new api
- updated the axios year patch route
- edited the server update year patch route
