+++ main/client/src/App.js
@@ -6,7 +6,8 @@ import { Route, Routes } from "react-router-dom";
 // We import all the components we need in our app
 import Navbar from "./components/nav/navbar";
 import CreateEnvelope from "./components/envelope/CreateEnvelope";
-import CreateYear from "./components/year/CreateYear1";
+import CreateYear from "./components/year/CreateYear";
+import EditYear from "./components/year/EditYear";
 import Years from "./components/year/Years";
 
 const App = () => {
@@ -16,6 +17,7 @@ const App = () => {
       <Routes>
         <Route exact path="/" element={<Years />} />
         <Route path="/createYear" element={<CreateYear type="Year" />} />
+        <Route path="/editYear/:year" element={<EditYear type="Year" />} />
         <Route path="/:year/:month/createEnvelope" element={<CreateEnvelope />} />
       </Routes>
     </div>
