+++ main/client/src/components/year/EditYear.js
@@ -1,12 +1,12 @@
 import React from "react";
-import YearForm from "./yearForm";
+import YearForm from "./YearForm";
  
 export default function EditYear() {
  
   // This following section will display the form that takes input from the user to update the data.
   return (
     <div>
-      <h3>Update Year</h3>
+      <h3>Edit Year</h3>
       <YearForm />
     </div>
   );
