+++ main/client/src/components/year/YearForm.js
@@ -1,12 +1,14 @@
 import React, { useState, useEffect } from "react";
 import { useNavigate, useParams } from "react-router-dom";
-import MonthsGrid from './MonthsGrid';
-import '../yearForm.css';
+import { generateMonthsArr } from "../../utils/general";
+import MonthsGrid from "./MonthsGrid";
+import api from "../../api/api";
+import "./year.css";
 
-export default function YearForm (props) {
+export default function YearForm(props) {
   const navigate = useNavigate();
   const params = useParams();
-  const [year, setYear] = useState("");
+  const [year, setYear] = useState(new Date().getFullYear());
   const [remaining, setRemaining] = useState(0);
   const [budget, setBudget] = useState("");
   const [distribution, setDistribution] = useState(0);
@@ -15,23 +17,7 @@ export default function YearForm (props) {
   useEffect(() => {
     if (params.year) {
       async function fetchData() {
-        const response = await fetch(
-          `http://localhost:5000/year/${params.year}`
-        );
-
-        if (!response.ok) {
-          const message = `An error has occurred: ${response.statusText}`;
-          window.alert(message);
-          return;
-        }
-
-        const record = await response.json();
-        if (!record) {
-          window.alert(`Record with id ${params.year} not found`);
-          navigate("/");
-          return;
-        }
-        const yearInfo = record[0];
+        const yearInfo = await api.getYear(params.year);
 
         setYear(yearInfo.year);
         setBudget(yearInfo.budget);
@@ -39,7 +25,12 @@ export default function YearForm (props) {
         const total = yearInfo.months.reduce((a, b) => a + b.budget, 0);
         const leftover = yearInfo.budget - total;
         setRemaining(leftover);
-        const currMonths = Object.fromEntries(yearInfo.months.map(month => [month.month, {budget: month.budget, max: month.budget + leftover}]));
+        const currMonths = Object.fromEntries(
+          yearInfo.months.map((month) => [
+            month.month,
+            { budget: month.budget, max: month.budget + leftover },
+          ])
+        );
         setMonths(currMonths);
       }
 
@@ -48,13 +39,13 @@ export default function YearForm (props) {
       return;
     }
   }, [params.year, navigate]);
-  
+
   const updateYear = (e) => setYear(Number(e.target.value));
 
   function getDistribution(n) {
     const distributed = Math.floor(n / 12);
     const remainder = n % 12;
-    return {distributed, remainder}
+    return { distributed, remainder };
   }
 
   function populateYears() {
@@ -63,16 +54,21 @@ export default function YearForm (props) {
     let yearsArr = [];
 
     for (let i = currentYear; i <= maxYear; i++) {
-      yearsArr.push((<option key={i} value={i}>{i}</option>));
+      yearsArr.push(
+        <option key={i} value={i}>
+          {i}
+        </option>
+      );
     }
 
     return yearsArr;
   }
 
   function createMonthsObject(distributed, remainder) {
-    const monthsArr = ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"];
-    
-    const months = Object.fromEntries(monthsArr.map(month => [month, {budget: distributed, max: distributed}]));
+    const monthsArr = generateMonthsArr();
+    const months = Object.fromEntries(
+      monthsArr.map((month) => [month, { budget: distributed, max: distributed }])
+    );
     months["January"].budget += remainder;
     months["January"].max += remainder;
     return months;
@@ -80,43 +76,25 @@ export default function YearForm (props) {
 
   const updateBudget = (e) => {
     const newBudget = e.target.value > 0 ? e.target.value : 0;
-    const {distributed, remainder} = getDistribution(newBudget);
-    setBudget(newBudget)
-    setRemaining(0)
-    setMonths(createMonthsObject(distributed, remainder))
+    const { distributed, remainder } = getDistribution(newBudget);
+    setBudget(newBudget);
+    setRemaining(0);
+    setMonths(createMonthsObject(distributed, remainder));
   };
 
   async function onSubmit(e) {
     e.preventDefault();
     const { isCreate } = props;
     const newYear = { year, budget, months };
-        
-    if (isCreate) {
-      // When a post request is sent to the create url, we'll add a new record to the database.
-      console.log(JSON.stringify(newYear));   
-      await fetch("http://localhost:5000/year/add", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(newYear),
-      })
-        .catch(error => {
-          window.alert(error);
-          return;
-        });
-    } else {
-      // This will send a patch request to update the data in the database.
-      await fetch(`http://localhost:5000/year/${year}`, {
-        method: "PATCH",
-        headers: {
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(newYear),
-      });
+    console.log("newYear", newYear);
+
+    try {
+      isCreate ? await api.createYear(newYear) : await api.updateYear(newYear);
+      navigate("/");
+    } catch(e) {
+      console.log(e);
+      window.alert(e);
     }
-        
-    navigate("/");
   }
 
   return (
@@ -125,12 +103,7 @@ export default function YearForm (props) {
         <form onSubmit={onSubmit}>
           <div className="form-group">
             <label htmlFor="year">Year</label>
-            <select
-              value={year}
-              className="form-control"
-              onChange={updateYear}
-              required
-            >
+            <select value={year} className="form-control" onChange={updateYear} required>
               {populateYears()}
             </select>
           </div>
@@ -147,7 +120,14 @@ export default function YearForm (props) {
             />
           </div>
           <div className="remaining">Remaining {remaining}</div>
-          <MonthsGrid distribution={distribution} budget={budget} months={months} setMonths={setMonths} remaining={remaining} setRemaining={setRemaining}/>
+          <MonthsGrid
+            distribution={distribution}
+            budget={budget}
+            months={months}
+            setMonths={setMonths}
+            remaining={remaining}
+            setRemaining={setRemaining}
+          />
           <div className="form-group">
             <input
               type="submit"
